Change iOS Linker settings in iOS Build
Changed it to "Link Framework SDKs Only" for both Debug and Release configurations. This is so that the latest version of Xcode needn't be installed on the development machine.

Also commented out some unneeded using statements.

Also made some async functions non-async because it wasn't necessary to make them async.